# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Biblioteca de Prompts API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserPlan = typeof UserPlan[keyof typeof UserPlan];


export const UserPlan = {
  basic: 'basic',
  pro: 'pro',
  premium: 'premium',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  plan: UserPlan;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  icon?: string | null;
  promptCount: number;
}

export interface CategoryInput {
  name: string;
  slug: string;
  icon?: string;
}

export type PromptVisibility = typeof PromptVisibility[keyof typeof PromptVisibility];


export const PromptVisibility = {
  public: 'public',
  private: 'private',
} as const;

export interface Prompt {
  id: number;
  title: string;
  content: string;
  /** @nullable */
  description?: string | null;
  categoryId: number;
  categoryName: string;
  tags?: string[];
  visibility: PromptVisibility;
  isFavorite: boolean;
  likeCount: number;
  useCount: number;
  rating: number;
  /** @nullable */
  authorName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type PromptInputVisibility = typeof PromptInputVisibility[keyof typeof PromptInputVisibility];


export const PromptInputVisibility = {
  public: 'public',
  private: 'private',
} as const;

export interface PromptInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  description?: string;
  categoryId: number;
  tags?: string[];
  visibility?: PromptInputVisibility;
}

export type PromptUpdateVisibility = typeof PromptUpdateVisibility[keyof typeof PromptUpdateVisibility];


export const PromptUpdateVisibility = {
  public: 'public',
  private: 'private',
} as const;

export interface PromptUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  content?: string;
  description?: string;
  categoryId?: number;
  tags?: string[];
  visibility?: PromptUpdateVisibility;
}

export interface PromptListResponse {
  prompts: Prompt[];
  total: number;
  page: number;
  limit: number;
}

export interface GenerateInput {
  /** @minLength 3 */
  topic: string;
  category?: string;
  style?: string;
}

export interface RefineInput {
  content: string;
  instruction: string;
}

export interface GenerateResponse {
  content: string;
  title: string;
  suggestedTags?: string[];
}

export interface UserStats {
  totalPrompts: number;
  favoritePrompts: number;
  publicPrompts: number;
  privatePrompts: number;
  totalUses: number;
  totalLikes: number;
  aiGenerationsUsed: number;
  aiGenerationsLimit: number;
  /** @nullable */
  promptsLimit?: number | null;
}

export type PlanSlug = typeof PlanSlug[keyof typeof PlanSlug];


export const PlanSlug = {
  basic: 'basic',
  pro: 'pro',
  premium: 'premium',
} as const;

export type PlanExclusiveTemplates = typeof PlanExclusiveTemplates[keyof typeof PlanExclusiveTemplates];


export const PlanExclusiveTemplates = {
  none: 'none',
  partial: 'partial',
  full: 'full',
} as const;

export interface Plan {
  id: number;
  name: string;
  slug: PlanSlug;
  price: number;
  /** @nullable */
  promptsLimit: number | null;
  aiGenerationsLimit: number;
  sharingEnabled: boolean;
  exclusiveTemplates: PlanExclusiveTemplates;
  description?: string;
}

export type SubscriptionStatus = typeof SubscriptionStatus[keyof typeof SubscriptionStatus];


export const SubscriptionStatus = {
  active: 'active',
  cancelled: 'cancelled',
  expired: 'expired',
} as const;

export interface Subscription {
  planSlug: string;
  planName: string;
  price: number;
  status: SubscriptionStatus;
  createdAt: string;
  /** @nullable */
  expiresAt?: string | null;
}

export type SubscriptionInputPlanSlug = typeof SubscriptionInputPlanSlug[keyof typeof SubscriptionInputPlanSlug];


export const SubscriptionInputPlanSlug = {
  basic: 'basic',
  pro: 'pro',
  premium: 'premium',
} as const;

export interface SubscriptionInput {
  planSlug: SubscriptionInputPlanSlug;
}

export type ListPromptsParams = {
category?: string;
search?: string;
visibility?: ListPromptsVisibility;
favorites?: string;
page?: number;
limit?: number;
};

export type ListPromptsVisibility = typeof ListPromptsVisibility[keyof typeof ListPromptsVisibility];


export const ListPromptsVisibility = {
  public: 'public',
  private: 'private',
  all: 'all',
} as const;

export type ListPopularPromptsParams = {
category?: string;
limit?: number;
};

export type ListRecentPromptsParams = {
limit?: number;
};