# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Biblioteca de Prompts API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Categories ─────────────────────────────────────────────────────────────
...
  /ai/generate:
    post:
      operationId: generatePrompt
      tags: [ai]
      summary: Generate a prompt with AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateInput"
      responses:
        "200":
          description: Generated prompt text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /ai/refine:
    post:
      operationId: refinePrompt
      tags: [ai]
      summary: Refine an existing prompt with AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefineInput"
      responses:
        "200":
          description: Refined prompt text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateResponse"

  # ─── Stats ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    User:
      type: object
      required: [id, name, email, plan, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        plan:
          type: string
          enum: [basic, pro, premium]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # ─── Category ─────────────────────────────────────────────────────────────
...
    GenerateInput:
      type: object
      required: [topic]
      properties:
        topic:
          type: string
          minLength: 3
        category:
          type: string
        style:
          type: string
    RefineInput:
      type: object
      required: [content, instruction]
      properties:
        content:
          type: string
        instruction:
          type: string
    GenerateResponse:
      type: object
      required: [content, title]
      properties:
        content:
          type: string
        title:
          type: string
        suggestedTags:
          type: array
          items:
            type: string

    # ─── Stats ─────────────────────────────────────────────────────────────────
...