# lib/api-zod/src/generated/types/generateInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Biblioteca de Prompts API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateInput {
  /** @minLength 3 */
  topic: string;
  category?: string;
  style?: string;
}