# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Biblioteca de Prompts API
 * OpenAPI spec version: 0.1.0
 */
import type { UserPlan } from './userPlan';

export interface User {
  id: number;
  name: string;
  email: string;
  plan: UserPlan;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}